Refactor preview and record helpers: move getPreviewFieldsConditions and checkShouldGroupWrap to preview.helper.ts; clean up Fields and ChildFields components
